Removed type safety for great good
In parse we had previously used a series of different data types to
represent the program at hand.  This created a great tree for us on its
own while parsing!

Despite that, we cannot create a function to traverse that tree, because
they were not all of the same type.

The implementation is now different; everything is a constructor for the
"cst" type.  This allows us to have traversal functions that rely on the
cst type, which is great.  However, we lost our type checking in the
constructors.  Now, Block (Close_Brace, Open_Brace, Plus) will
technically work.  We must take greater caution in the programs
generating the cst, to ensure the correct relations are kept.

Luckily - because the program was written and worked already with
everything as a separate type - we are already in that state, providing
we never create new elements for that cst